refactored the code, created the decryptFiles function

+++ Stockholm/src/CryptoManager.cpp
@@ -3,7 +3,6 @@
 #include <iostream>
 #include <fstream>
 
-#include <iomanip>
 
 cryptoManager::~cryptoManager()
 {
@@ -29,6 +28,13 @@ void cryptoManager::writeEncryptSymmKey()
     _out.write(reinterpret_cast<const char *>(_symKeyEncrypted.data()), _symKeyEncrypted.size());
 }
 
+bool cryptoManager::initSecretStreamPush()
+{
+    if (crypto_secretstream_xchacha20poly1305_init_push(&_state, _header.data(), _symKey.data()) != 0)
+        return (_logs.printError("Failed to initialize XChaCha20-Poly1305 stream"), false);
+    return true;
+}
+
 void cryptoManager::writeHeader()
 {
     _out.write(reinterpret_cast<const char *>(_header.data()), _header.size());
@@ -41,13 +47,15 @@ void cryptoManager::writeEncryptedData()
         _in.read(reinterpret_cast<char *>(_buffer.data()), _buffer.size());
         std::streamsize bytesRead = _in.gcount();
         if (bytesRead > 0)
-        {
-            unsigned long long clen;
-            _encryptedBuffer.resize(bytesRead + STREAM_MAC_SIZE);
-            crypto_secretstream_xchacha20poly1305_push(&_state, _encryptedBuffer.data(), &clen, _buffer.data(), bytesRead, nullptr, 0, 0);
-            _out.write(reinterpret_cast<char *>(_encryptedBuffer.data()), clen);
-        }
+            cryptAndWriteData(bytesRead);
     }
+}
+void cryptoManager::cryptAndWriteData(std::streamsize bytesRead)
+{
+    unsigned long long clen;
+    _encryptedBuffer.resize(bytesRead + STREAM_MAC_SIZE);
+    crypto_secretstream_xchacha20poly1305_push(&_state, _encryptedBuffer.data(), &clen, _buffer.data(), bytesRead, nullptr, 0, 0);
+    _out.write(reinterpret_cast<char *>(_encryptedBuffer.data()), clen);
 }
 
 void cryptoManager::init()
@@ -73,7 +81,8 @@ void cryptoManager::encryptFiles()
         generateSymKey();
         encryptSymmKey();
         writeEncryptSymmKey();
-        crypto_secretstream_xchacha20poly1305_init_push(&_state, _header.data(), _symKey.data());
+        if (!initSecretStreamPush())
+            continue;
         writeHeader();
         writeEncryptedData();
         fs::remove(file);
@@ -88,91 +97,89 @@ bool cryptoManager::openFileStreams(fs::path file)
         return (_logs.printError("Failed to open: " + file.filename().string()), false);
     _out = _filemanager.createNewFile(file, _tmpPath);
     if (!_out)
-        return (_logs.printError("Creation of the encrypted file failed for file: " + file.filename().string()), false);
+        return (_logs.printError("Creation of the new file failed for file: " + file.filename().string()), false);
     return true;
 }
 
 void cryptoManager::decryptFiles()
 {
-
-    hexToBinPrivKey();
+    hexToBinAsymPrivKey();
+    hexToBinAsymPubKey();
     for (auto &file : _files)
     {
-        _in = _filemanager.openOriginalFile(file);
-        if (!_in)
-        {
-            _logs.printError("Failed to open: " + file.filename().string());
+        if (!openFileStreams(file))
             continue;
-        }
-
-        _out = _filemanager.createNewFile(file, _tmpPath);
-        if (!_out)
-        {
-            _logs.printError("Creation of the encrypted file failed for file: " + file.filename().string());
+        readEncryptedSymKey();
+        if (!decryptSymKey())
             continue;
-        }
-        _in.read(reinterpret_cast<char *>(_symKeyEncrypted.data()), _symKeyEncrypted.size());
-        hexToBinAsymPubKey();
-
-        if (crypto_box_seal_open(
-                _symKey.data(),
-                _symKeyEncrypted.data(),
-                _symKeyEncrypted.size(),
-                _asymKeyPub.data(),
-                _asymKeyPriv.data()) != 0)
-        {
-            _logs.printError("Failed to decrypt symmetric key");
+        readSecretStreamHeader();
+        if (!initSecretStreamPull())
             continue;
-        }
-
-        // convertsymKeyToHex();
-        // std::cout << "sym key = " << _hex_symKey.data() << std::endl;
-        _in.read(reinterpret_cast<char *>(_header.data()), _header.size());
-        if (crypto_secretstream_xchacha20poly1305_init_pull(
-                &_state,
-                _header.data(),
-                _symKey.data()) != 0)
-        {
-
-            _logs.printError("Invalid key or header");
-        }
-        _encryptedBuffer.resize(ENCRYPTED_BUFFER_SIZE);
-        unsigned char tag;
-        unsigned long long out_len;
-
-        while (_in)
-        {
-            _in.read(reinterpret_cast<char *>(_encryptedBuffer.data()), _encryptedBuffer.size());
-            std::streamsize bytesRead = _in.gcount();
-            if (bytesRead > 0)
-            {
-                if (crypto_secretstream_xchacha20poly1305_pull(
-                        &_state,
-                        _buffer.data(),
-                        &out_len,
-                        &tag,
-                        _encryptedBuffer.data(),
-                        bytesRead,
-                        nullptr,
-                        0) != 0)
-                {
-                    _logs.printError("Data decryption failed");
-                }
-            }
-            _out.write(reinterpret_cast<char *>(_buffer.data()), out_len);
-            if (tag == TAG_FINAL)
-                break;
-        }
+        writeDecryptedData();
         fs::remove(file);
         _filemanager.removeFtExt(_tmpPath);
     }
 }
 
-//------------------------------------------------------------------------
-
-void cryptoManager::hexToBinPrivKey()
+void cryptoManager::readEncryptedSymKey()
 {
-    // sodium_hex2bin(_asymKeyPriv.data(), _asymKeyPriv.size(), _conf.reverseKey.c_str(), _conf.reverseKey.size(), nullptr, nullptr, nullptr);
+    _in.read(reinterpret_cast<char *>(_symKeyEncrypted.data()), _symKeyEncrypted.size());
+}
+bool cryptoManager::decryptSymKey()
+{
+    if (crypto_box_seal_open(_symKey.data(), _symKeyEncrypted.data(), _symKeyEncrypted.size(),
+                             _asymKeyPub.data(), _asymKeyPriv.data()) != 0)
+        return (_logs.printError("Failed to decrypt symmetric key"), false);
+    return true;
+}
+
+void cryptoManager::readSecretStreamHeader()
+{
+    _in.read(reinterpret_cast<char *>(_header.data()), _header.size());
+}
+
+bool cryptoManager::initSecretStreamPull()
+{
+    if (crypto_secretstream_xchacha20poly1305_init_pull(&_state, _header.data(), _symKey.data()) != 0)
+        return (_logs.printError("Invalid key or header"), false);
+    return true;
+}
+
+void cryptoManager::writeDecryptedData()
+{
+    _encryptedBuffer.resize(ENCRYPTED_BUFFER_SIZE);
+    unsigned char tag;
+    unsigned long long out_len;
+
+    while (_in)
+    {
+        readCryptedData();
+        std::streamsize bytesRead = _in.gcount();
+        if (bytesRead > 0 && !decryptData(tag, out_len, bytesRead))
+            break;
+        _out.write(reinterpret_cast<char *>(_buffer.data()), out_len);
+        if (tag == TAG_FINAL)
+            break;
+    }
+}
+
+void cryptoManager::readCryptedData()
+{
+    _in.read(reinterpret_cast<char *>(_encryptedBuffer.data()), _encryptedBuffer.size());
+}
+
+bool cryptoManager::decryptData(unsigned char &tag, unsigned long long &out_len, std::streamsize bytesRead)
+{
+    if (crypto_secretstream_xchacha20poly1305_pull(&_state, _buffer.data(), &out_len,
+                                                   &tag, _encryptedBuffer.data(), bytesRead, nullptr, 0) != 0)
+        return (_logs.printError("Data decryption failed"), false);
+    return true;
+}
+
+
+void cryptoManager::hexToBinAsymPrivKey()
+{
+  
     if (sodium_hex2bin(
             _asymKeyPriv.data(),
             _asymKeyPriv.size(),
